Rename folder to use JS convention
