+++ backend/src/lib/event-logger/index.js
