+++ backend/src/lib/event-logger/index.spec.js
